1st draft of refactor to package
