--- cmake/build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-./configure --prefix=$PREFIX
-make
-make install
