--- cmake/meta.yaml
@@ -1,22 +1,0 @@
-package:
-  name: cmake
-  version: 3.3.1
-
-source:
-  url: http://www.cmake.org/files/v3.3/cmake-3.3.1.tar.gz  [not win]
-  fn: cmake-3.3.1.tar.gz  [not win]
-  url: http://www.cmake.org/files/v3.3/cmake-3.3.1-win32-x86.zip [win]
-  fn: cmake-3.3.1-win32-x86.zip [win]
-
-build:
-  binary_relocation: False           [osx]
-
-test:
-  commands:
-    - cmake --version [not win]
-
-about:
-  home: http://www.cmake.org/
-  license:  3-clause BSD
-  license_family: BSD
-  summary: CMake is an extensible, open-source system that manages the build process
