--- cmake/run_test.bat
@@ -1,2 +1,0 @@
-set PATH=%LIBRARY_BIN%;%PATH%
-cmake --version
