--- cython/meta.yaml
@@ -1,33 +1,0 @@
-package:
-  name: cython
-  version: 0.23.4
-
-source:
-  fn: Cython-0.23.4.tar.gz
-  url: https://pypi.python.org/packages/source/C/Cython/Cython-0.23.4.tar.gz
-  md5: 157df1f69bcec6b56fd97e0f2e057f6e
-
-build:
-  number: 1           [osx]
-  entry_points:
-    - cython = Cython.Compiler.Main:setuptools_main
-    - cythonize = Cython.Build.Cythonize:main
-    - cygdb = Cython.Debugger.Cygdb:main
-
-requirements:
-  build:
-    - python
-    - setuptools       [win]
-  run:
-    - python
-
-test:
-  commands:
-    - cython --version
-  files:
-    - fib.pyx
-
-about:
-  home: http://www.cython.org/
-  license: Apache
-  summary: The Cython compiler for writing C extensions for the Python language
