--- hdf5/vs_2015_cmake.patch
@@ -1,88 +1,0 @@
-diff --git config/cmake_ext_mod/ConfigureChecks.cmake config/cmake_ext_mod/ConfigureChecks.cmake
-index 588b5d5..b5e5c34 100644
---- config/cmake_ext_mod/ConfigureChecks.cmake
-+++ config/cmake_ext_mod/ConfigureChecks.cmake
-@@ -90,7 +90,9 @@ if (WINDOWS)
-   endif (NOT UNIX AND NOT CYGWIN AND NOT MINGW)
-   set (${HDF_PREFIX}_HAVE_FUNCTION 1)
-   set (${HDF_PREFIX}_GETTIMEOFDAY_GIVES_TZ 1)
--  set (${HDF_PREFIX}_HAVE_TIMEZONE 1)
-+  if (MSVC_VERSION LESS 1900)
-+    set (${HDF_PREFIX}_HAVE_TIMEZONE 1)
-+  endif (MSVC_VERSION LESS 1900)
-   set (${HDF_PREFIX}_HAVE_GETTIMEOFDAY 1)
-   if (MINGW)
-     set (${HDF_PREFIX}_HAVE_WINSOCK2_H 1)
-@@ -718,35 +720,33 @@ endif (INLINE_TEST___inline__)
- if (NOT ${HDF_PREFIX}_PRINTF_LL_WIDTH OR ${HDF_PREFIX}_PRINTF_LL_WIDTH MATCHES "unknown")
-   set (PRINT_LL_FOUND 0)
-   message (STATUS "Checking for appropriate format for 64 bit long:")
--  foreach (HDF5_PRINTF_LL l64 l L q I64 ll)
--    set (CURRENT_TEST_DEFINITIONS "-DPRINTF_LL_WIDTH=${HDF5_PRINTF_LL}")
--    if (${HDF_PREFIX}_SIZEOF_LONG_LONG)
--      set (CURRENT_TEST_DEFINITIONS "${CURRENT_TEST_DEFINITIONS} -DHAVE_LONG_LONG")
--    endif (${HDF_PREFIX}_SIZEOF_LONG_LONG)
--    TRY_RUN (HDF5_PRINTF_LL_TEST_RUN   HDF5_PRINTF_LL_TEST_COMPILE
--        ${CMAKE_BINARY_DIR}
--        ${HDF_RESOURCES_EXT_DIR}/HDFTests.c
--        CMAKE_FLAGS -DCOMPILE_DEFINITIONS:STRING=${CURRENT_TEST_DEFINITIONS}
--        OUTPUT_VARIABLE OUTPUT
--    )
--    if (HDF5_PRINTF_LL_TEST_COMPILE)
--      if (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
--        set (${HDF_PREFIX}_PRINTF_LL_WIDTH "\"${HDF5_PRINTF_LL}\"" CACHE INTERNAL "Width for printf for type `long long' or `__int64', us. `ll")
--        set (PRINT_LL_FOUND 1)
--      else (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
--        message ("Width with ${HDF5_PRINTF_LL} failed with result: ${HDF5_PRINTF_LL_TEST_RUN}")
--      endif (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
--    else (HDF5_PRINTF_LL_TEST_COMPILE)
--      file (APPEND ${CMAKE_BINARY_DIR}/CMakeFiles/CMakeError.log
--          "Test ${HDF_PREFIX}_PRINTF_LL_WIDTH for ${HDF5_PRINTF_LL} failed with the following output:\n ${OUTPUT}\n"
--      )
--    endif (HDF5_PRINTF_LL_TEST_COMPILE)
--  endforeach (HDF5_PRINTF_LL)
-+  set (CURRENT_TEST_DEFINITIONS "-DPRINTF_LL_WIDTH")
-+  if (H5_SIZEOF_LONG_LONG)
-+    set (CURRENT_TEST_DEFINITIONS "${CURRENT_TEST_DEFINITIONS} -DHAVE_LONG_LONG")
-+  endif (H5_SIZEOF_LONG_LONG)
-+  TRY_RUN (HDF5_PRINTF_LL_TEST_RUN   HDF5_PRINTF_LL_TEST_COMPILE
-+      ${HDF5_BINARY_DIR}/CMake
-+      ${HDF_RESOURCES_EXT_DIR}/HDFTests.c
-+      CMAKE_FLAGS -DCOMPILE_DEFINITIONS:STRING=${CURRENT_TEST_DEFINITIONS}
-+      OUTPUT_VARIABLE OUTPUT
-+  )
-+  if (HDF5_PRINTF_LL_TEST_COMPILE)
-+    if (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
-+      string(REGEX REPLACE ".*PRINTF_LL_WIDTH=\\[(.*)\\].*" "\\1" HDF5_PRINTF_LL "${OUTPUT}")
-+      set (H5_PRINTF_LL_WIDTH "\"${HDF5_PRINTF_LL}\"" CACHE INTERNAL "Width for printf for type `long long' or `__int64', us. `ll")
-+      set (PRINT_LL_FOUND 1)
-+    else (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
-+      message ("HDF5: Width test failed with result: ${HDF5_PRINTF_LL_TEST_RUN}")
-+    endif (HDF5_PRINTF_LL_TEST_RUN MATCHES 0)
-+  else (HDF5_PRINTF_LL_TEST_COMPILE)
-+    file (APPEND ${CMAKE_BINARY_DIR}/CMakeFiles/CMakeError.log
-+        "Test H5_PRINTF_LL_WIDTH failed with the following output:\n ${OUTPUT}\n")
-+  endif (HDF5_PRINTF_LL_TEST_COMPILE)
- 
-   if (PRINT_LL_FOUND)
--    message (STATUS "Checking for apropriate format for 64 bit long: found ${${HDF_PREFIX}_PRINTF_LL_WIDTH}")
-+    message (STATUS "Checking for appropriate format for 64 bit long: found ${${HDF_PREFIX}_PRINTF_LL_WIDTH}")
-   else (PRINT_LL_FOUND)
--    message (STATUS "Checking for apropriate format for 64 bit long: not found")
-+    message (STATUS "Checking for appropriate format for 64 bit long: not found")
-     set (${HDF_PREFIX}_PRINTF_LL_WIDTH "\"unknown\"" CACHE INTERNAL
-         "Width for printf for type `long long' or `__int64', us. `ll"
-     )
-diff --git config/cmake_ext_mod/HDFTests.c config/cmake_ext_mod/HDFTests.c
-index 1c23ac3..5c4d360 100644
---- config/cmake_ext_mod/HDFTests.c
-+++ config/cmake_ext_mod/HDFTests.c
-@@ -213,7 +213,7 @@ SIMPLE_TEST(struct stat sb; sb.st_blocks=0);
- 
- int main(void)
- {
--  char *llwidthArgs[] = { "l64", "l", "L", "q", "ll", NULL };
-+  char *llwidthArgs[] = { "I64", "l64", "l", "L", "q", "ll", NULL };
-   char *s = malloc(128);
-   char **currentArg = NULL;
-   LL_TYPE x = (LL_TYPE)1048576 * (LL_TYPE)1048576;
