+++ meta.yaml
@@ -1,0 +1,19 @@
+package:
+  name: protoci
+  version: 0.0.0
+
+source:
+  git_rev: 0.0.0
+  git_url: https://github.com/ContinuumIO/ProtoCI
+
+build:
+  number: 1
+
+requirements:
+  build:
+    - python
+    - setuptools
+  run:
+    - python
+about:
+  summary: Helps with building conda packages in anaconda-build.
