+++ protoci/build2.py
@@ -14,7 +14,6 @@ import networkx as nx
 import sys
 
 
-
 from conda_build.metadata import parse, MetaData
 
 CONDA_BUILD_CACHE=os.environ.get("CONDA_BUILD_CACHE")
@@ -158,51 +157,6 @@ def construct_graph(directory):
 
     return g
 
-def successors_iter(g, s, nodes):
-    for s in g.successors(s):
-        nodes.add(s)
-        for s in tuple(successors_iter(g, s, nodes)):
-            nodes.add(s)
-    return nodes
-
-def coalesce(hi_level_builds, targetnum):
-    coalesced = defaultdict(lambda: [])
-    counts = [(k, len(v)) for k, v in hi_level_builds.items()]
-    group = []
-    for key, count in sorted(counts, key=lambda x:x[1]):
-        group.append(hi_level_builds[key] + [key])
-        if sum(map(len, group)) >= targetnum:
-            for g in group:
-                if g == key:
-                    continue
-                coalesced[key] += [gi for gi in g if gi not in coalesced[key] and gi != key]
-            group = []
-    if group:
-        for g in group:
-            coalesced[key] += [gi for gi in g if gi not in coalesced[key] and gi != key]
-    return coalesced
-
-def split_graph(g, targetnum, split_file):
-    g = g.copy()
-    toposort = nx.topological_sort(g)
-    packages_covered = defaultdict(lambda:0)
-    degrees = dict(g.degree_iter())
-
-    hi_level_builds = {}
-    for hi_level in nx.topological_sort(g):
-        if hi_level in packages_covered:
-            continue
-        succ = tuple(successors_iter(g, hi_level, set()))
-        for s in succ:
-            packages_covered[s] += 1
-        packages_covered[hi_level] += 1
-        topo_order = [(s, toposort.index(s)) for s in succ]
-        succ_order = sorted(topo_order, key=lambda x: -x[1])
-        hi_level_builds[hi_level] = [_[0] for _ in succ_order]
-    hi_level_builds = coalesce(hi_level_builds, targetnum)
-    with open(split_file, 'w') as f:
-        f.write(json.dumps(hi_level_builds))
-    return hi_level_builds
 
 def build_order(graph, packages, level=0):
     '''
@@ -290,207 +244,34 @@ def make_pkg(package, dry=False, extra_args=''):
             raise
 
 
-def submit_one(args):
-    '''
-    Adjusts binstar_template.yml
-    base on
-        user
-        queue
-        build arguments
-        platforms
-    Comes up with a package name for user
-    Creates package if it doesn't exist
-    Submits package
-    Prints out the command you need to tail the build
-    returns 0 if okay
-    '''
-    import jinja2
-    js_file, key = args.json_file_key
-    with open(js_file, 'r') as f:
-        js = json.load(f)
-    with open(os.path.join(os.path.dirname(__file__), 'binstar_template.yml')) as f:
-        contents = f.read()
-        t = jinja2.Template(contents)
-        package = 'protoci-' + key
-        info = (os.path.basename(js_file), key)
-        platforms = "".join(" - {}\n".format(p) for p in args.platforms)
-        binstar_yml = t.render(PACKAGE=package,
-                               USER=args.user,
-                               PLATFORMS=platforms,
-                               BUILD_ARGS='./ build ' +\
-                                          '-json-file-key {0} {1}'.format(*info))
-        with open(os.path.join(args.path, '.binstar.yml'), 'w') as f:
-            f.write(binstar_yml)
-    full_package = '{0}/{1}'.format(args.user, package)
-    cmd = ['anaconda', 'build', 'list-all', full_package]
-    print('Check to see if', full_package, 'exists:', cmd)
-    proc = subprocess.Popen(cmd)
-    if proc.wait():
-        cmd = ['anaconda', 'package','--create', full_package]
-        print("prepare to create package", cmd)
-        if not args.dry:
-            create = subprocess.Popen(cmd, cwd=args.path)
-            if create.wait():
-                raise ValueError('Could not create {}'.format(full_package))
-
-    user_queue = '{0}/{1}'.format(args.user, args.queue)
-    cmd = ['anaconda', 'build',
-           'submit', './', '--queue',
-           user_queue]
-    print('prepare to submit', cmd)
-    if args.dry:
-        return 0
-    proc =  subprocess.Popen(cmd, cwd=args.path, stdout=subprocess.PIPE,
-                             stderr=subprocess.PIPE)
-    ret = proc.wait()
-    out = proc.stdout.read().decode()
-    tail = [line for line in out.split('\n')
-            if 'tail' in line and full_package in line]
-    if len(tail):
-        tail = tail[0]
-    else:
-        print("Apparently something wrong with:", out)
-        time.sleep(10)
-    print('TAIL:\t', tail)
-    return ret
-
-
-def submit_full_json(args):
-    ''' Given -full-json, run every package tree
-    in a json that was created by split action, typically
-    called package_tree.js
-    '''
-    with open(args.full_json, 'r') as f:
-        tree = json.load(f)
-        print('{} high level packages'.format(len(tree)))
-        print('\twith total packages:',
-              len(tree) + sum(map(len, tree.values())))
-        for key in tree:
-            print('Key: ', key, len(tree[key])+1, 'packages to build/test')
-            args.json_file_key = (args.full_json, key)
-            submit_one(args)
-    return 0
-
-def pre_submit_clean_up(args):
-    '''Copies files from patterns like:
-
-    ./special_cases/<package-name>/run_test.sh
-
-    to
-
-    args.path/<package-name>/run_test.sh
-
-    (Helpful if anaconda-build needs mods)
-    '''
-    special = os.path.join(os.path.dirname(__file__), 'special_cases')
-    for dirr in os.listdir(special):
-        for fil in os.listdir(os.path.join(special, dirr)):
-            full_file = os.path.join(special, dirr, fil)
-            if not os.path.exists(os.path.join(args.path, dirr)):
-                continue
-            target = os.path.join(args.path, dirr, fil)
-            print('Copy', full_file, 'to', target)
-            print('Copy', full_file, 'to', target+'_removed')
-            shutil.copy(full_file, target + '_removed')
-            shutil.copy(full_file, target)
-
-    this_file = os.path.basename(__file__)
-    build2_in_other_dir = os.path.abspath(os.path.join(args.path, this_file))
-    shutil.copy(__file__, build2_in_other_dir)
-    print('Copy',__file__,'to', build2_in_other_dir)
-    package_tree_file = os.path.abspath(args.full_json or args.json_file_key[0])
-    n = datetime.datetime.now()
-    datestr = "_".join(map(str, (n.year, n.month, n.day, n.hour, n.minute, n.second)))
-    branch_name = 'build_' + datestr
-    print('Make a scratch git branch in', os.path.abspath(args.path))
-    subprocess.check_output(['git', 'checkout', '-b', branch_name], cwd=args.path)
-    subprocess.check_output(['git', 'add', build2_in_other_dir, package_tree_file], cwd=args.path)
-    print(subprocess.Popen(
-          ['git', 'commit', '-m',
-          'commit build2.py and the package json for anaconda-build'],
-          cwd=args.path).communicate())
-
-
-def submit_helper(args):
-    pre_submit_clean_up(args)
-    if 'submit' in sys.argv:
-        if args.full_json:
-            return submit_full_json(args)
-        else:
-            assert len(args.json_file_key) >= 2
-            arg1 = args.json_file_key[0]
-            hi_level = args.json_file_key[1:]
-            for key in hi_level:
-                args.json_file_key = (arg1, key)
-                ret_val = submit_one(args)
-                if ret_val:
-                    return ret_val
-    return 0
-
-
-
-def cli(parse_this=None):
-    p = argparse.ArgumentParser()
-    p.add_argument("path", default='.')
-    subp = p.add_subparsers(help="Build or split to make json of package "
-                                 "order/grouping. \n\tChoices: %(choices)s")
-    build_parser = subp.add_parser('build')
-    build_pkgs = build_parser.add_mutually_exclusive_group()
+def sequential_build_cli(parse_this=None):
+    parser = argparse.ArgumentParser()
+    parser.add_argument("path", default='.')
+    build_pkgs = parser.add_mutually_exclusive_group()
     build_pkgs.add_argument("-build", action='append', default=[])
     build_pkgs.add_argument("-buildall", action='store_true')
     build_pkgs.add_argument('-json-file-key', default=[], nargs="+",
                             help="Example: -json-file-key package_tree.js libnetcdf pysam")
-    build_parser.add_argument("-dry", action='store_true', default=False,
+    parser.add_argument("-dry", action='store_true', default=False,
                               help="Dry run")
-    build_parser.add_argument("-api", action='store_true', dest='recompile',
+    parser.add_argument("-api", action='store_true', dest='recompile',
                               default=False)
-    build_parser.add_argument("-args", action='store', dest='cbargs', default='')
-    build_parser.add_argument("-l", type=int, action='store', dest='level', default=0)
-    build_parser.add_argument("-noautofail", action='store_false', dest='autofail', default=True)
-    split_parser = subp.add_parser('split')
-    split_parser.add_argument('-t','--targetnum', type=int,
-                              default=10,
-                              help="How many packages in one anaconda "
-                                   "build submission typically.")
-    split_parser.add_argument('-s','--split-files',type=str,default="package_tree.js")
-    submit_parser = subp.add_parser('submit')
-    json_read_choice = submit_parser.add_mutually_exclusive_group()
-    json_read_choice.add_argument('-json-file-key',
-                                  default=[], nargs=2,
-                                help="Example: -json-file-key package_tree.js libnetcdf")
-    json_read_choice.add_argument('-full-json',
-                                  type=str,
-                                  help="Build all packages named in json of splits")
-    submit_parser.add_argument('-user', default='conda-team',
-                               help="Anaconda username. Default: %(default)s")
-    submit_parser.add_argument('-queue', default='build_recipes',
-                               help="Anaconda build queue. Default: %(default)s")
-    submit_parser.add_argument('-dry', action='store_true', help='Dry run')
-    submit_parser.add_argument('-platforms', required=True,
-                               help="Some of all of %(default)s",
-                               default=['osx-64', 'linux-64','win-64'],
-                               nargs="+")
+    parser.add_argument("-args", action='store', dest='cbargs', default='')
+    parser.add_argument("-l", type=int, action='store', dest='level', default=0)
+    parser.add_argument("-noautofail", action='store_false', dest='autofail', default=True)
     if parse_this is None:
-        args = p.parse_args()
+        args = parser.parse_args()
     else:
-        args = p.parse_args(parse_this)
+        args = parser.parse_args(parse_this)
     print('Running build2.py with args of', args)
     if getattr(args, 'json_file_key', None):
         assert len(args.json_file_key) == 2, 'Should be 2 args: json_filename key'
     return args
 
-if __name__ == "__main__":
 
-    args = cli()
-    print("%s" % (getattr(args,'build','')))
-    print("-------------------------------")
-    if 'submit' in sys.argv:
-        sys.exit(submit_helper(args))
+def sequential_build_main(parse_this=None):
+    args = sequential_build_cli(parse_this=parse_this)
     g = construct_graph(args.path)
-    if getattr(args, 'split_files', None) is not None:
-        split_graph(g, args.targetnum, args.split_files)
-        print("See ", args.split_files, 'for split packages')
-        sys.exit(0)
     try:
         if args.buildall:
             args.build = None
@@ -508,7 +289,6 @@ if __name__ == "__main__":
                                          extra_args=args.cbargs,
                                          level=args.level,
                                          autofail=args.autofail)
-
         print("BUILD SUMMARY:")
         print("SUCCESS: [{}]".format(', '.join(success)))
         print("FAIL: [{}]".format(', '.join(fail)))
@@ -527,3 +307,16 @@ if __name__ == "__main__":
         sys.exit(len(fail))
     except:
         raise
+
+
+def difference_build_cli(parse_this=None):
+    parser = argparse.ArgumentParser(description="Does git diff to determine build order.")
+    # add arguments here related to git diff options
+    if not parse_this:
+        return parser.parse_args()
+    return parser.parse_args(parse_this)
+
+def difference_build_main(parse_this=None):
+    args = difference_build_cli(parse_this=parse_this)
+    print('Make differencing of git function here.')
+    return
