--- pysam/Gamma.h
@@ -1,19 +1,0 @@
-// Visit http://www.johndcook.com/stand_alone_code.html for the source of this code and more like it.
-
-#ifndef GAMMA_H
-#define GAMMA_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-// Note that the functions Gamma and LogGamma are mutually dependent.
-double LogGamma(double);
-double Gamma(double);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
-
