--- pysam/build.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-if [ `uname` == Linux ]; then
-    export CFLAGS="-I$PREFIX/include -L$PREFIX/lib"
-fi
-
-rm ez_setup.*
-
-$PYTHON setup.py install
