--- pysam/meta.yaml
@@ -1,23 +1,0 @@
-package:
-  name: pysam
-  version: 0.6
-
-source:
-  fn: pysam-0.6.tar.gz
-  url: http://pysam.googlecode.com/files/pysam-0.6.tar.gz
-  md5: 395f59d7b765d9f625f6d82fce905dc7
-  patches:
-    - no_setuptools.patch
-    - pysam_windows.patch  [win]
-
-requirements:
-  build:
-    - cython
-    - zlib
-  run:
-    - python
-
-about:
-  home: http://code.google.com/p/pysam/
-  license: MIT
-  summary: python module for reading and manipulating Samfiles
