--- pysam/no_setuptools.patch
@@ -1,15 +1,0 @@
---- setup.py
-+++ setup.py
-@@ -114,11 +114,7 @@ if len(sys.argv) >= 2 and sys.argv[1] == "refresh":
-     sys.exit(0)
- 
- 
--
--from ez_setup import use_setuptools
--use_setuptools()
--
--from setuptools import Extension, setup
-+from distutils.core import Extension, setup
- 
- ## note that for automatic cythoning, 
- ## both pyrex and cython need to be installed.
