--- pysam/run_test.py
@@ -1,4 +1,0 @@
-import TabProxies
-import csamtools
-import ctabix
-# import cvcf   FIXME
