--- pysam/stdint.h.win32
@@ -1,22 +1,0 @@
-#ifndef STDINT_H
-#define STDINT_H
-
-typedef signed char      int8_t;
-typedef short            int16_t;
-typedef int              int32_t;
-typedef __int64          int64_t;
-typedef ptrdiff_t        intptr_t;
-typedef unsigned char    uint8_t;
-typedef unsigned short   uint16_t;
-typedef unsigned int     uint32_t;
-typedef unsigned __int64 uint64_t;
-typedef size_t           uintptr_t;
-
-// A hack because pysam uses "inline" in C code
-#ifndef __cplusplus
-#define inline __inline
-#endif
-
-
-
-#endif // STDINT_H
