--- zlib/build.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-./configure --shared --prefix=$PREFIX
-make
-make install
-
-rm -rf $PREFIX/share
