--- zlib/meta.yaml
@@ -1,25 +1,0 @@
-package:
-  name: zlib
-  version: 1.2.8
-
-source:
-  fn: zlib-1.2.8.tar.gz
-  url: http://zlib.net/zlib-1.2.8.tar.gz
-  md5: 44d667c142d7cda120332623eab69f40
-
-build:
-  number: 2   [win]
-  features:
-    - vc9     [win and py27]
-    - vc10    [win and (py33 or py34)]
-    - vc14    [win and py35]
-
-requirements:
-  build:
-    - cmake
-
-about:
-  home: http://zlib.net/
-  license: zlib (http://zlib.net/zlib_license.html)
-  license_family: Other
-  summary: massively spiffy yet delicately unobtrusive compression library
